Add MB85RS1MT to the supported devices list
Hello,

Thank you for the library. I added the 1MBit Fujitsu FRAM module to the supported devices list. I also changed the line that prints Flash Size into the FRAM size since this is not a Flash device. I think this is still there because it might be ported from a Flash IC library.